refacto: extract classes into dedicated module
